cdk/requirements.txt
aws-cdk-lib==2.21.0
constructs>=10.0.0,<11.0.0
git+https://github.com/IGVF-DACC/shared-infrastructure@main#egg=shared_infrastructure
